Applied reviewer suggestions on template dev_container

+++ startup.sh
@@ -4,8 +4,10 @@
 # ----------------------------------------------------------------------
 # infDB Setup Script
 # ----------------------------------------------------------------------
-echo "=== Copy config-infdb.yml from template ==="
-cp configs/config-infdb.yml.template configs/config-infdb.yml
+if [ ! -f configs/config-infdb.yml ]; then
+    echo "=== Copy config-infdb.yml from template ==="
+    cp configs/config-infdb.yml.template configs/config-infdb.yml
+fi
 
 docker compose -f services/infdb-setup/compose.yml up
